# frontend/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost","description":"Generated server url"}],"paths":{"/api/v1/products/{id}":{"get":{"tags":["product-controller"],"operationId":"getProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}},"put":{"tags":["product-controller"],"operationId":"updateProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}}},"/api/v1/policies/{policyId}/commission":{"put":{"tags":["policy-controller"],"operationId":"configureCommission","parameters":[{"name":"policyId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfigureCommissionRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/customers/{id}":{"get":{"tags":["customer-controller"],"operationId":"getCustomer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CustomerResponse"}}}}}},"put":{"tags":["customer-controller"],"operationId":"updateCustomer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCustomerRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CustomerResponse"}}}}}}},"/api/v1/products":{"get":{"tags":["product-controller"],"operationId":"listProducts","parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string"}},{"name":"category","in":"query","required":false,"schema":{"type":"string"}},{"name":"status","in":"query","required":false,"schema":{"type":"string"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageProductResponse"}}}}}},"post":{"tags":["product-controller"],"operationId":"createProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}}},"/api/v1/products/{id}/lifecycle/deactivate":{"post":{"tags":["product-controller"],"operationId":"deactivateProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/products/{id}/lifecycle/activate":{"post":{"tags":["product-controller"],"operationId":"activateProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/policies":{"get":{"tags":["policy-controller"],"operationId":"listPolicies","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagePolicyResponse"}}}}}},"post":{"tags":["policy-controller"],"operationId":"createPolicy","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePolicyRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PolicyResponse"}}}}}}},"/api/v1/policies/{policyId}/lifecycle/reactivate":{"post":{"tags":["policy-controller"],"operationId":"reactivatePolicy","parameters":[{"name":"policyId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LifecycleRequest"}}}},"responses":{"200":{"description":"OK"}}}},"/api/v1/policies/{policyId}/lifecycle/deactivate":{"post":{"tags":["policy-controller"],"operationId":"deactivatePolicy","parameters":[{"name":"policyId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/policies/{policyId}/lifecycle/activate":{"post":{"tags":["policy-controller"],"operationId":"activatePolicy","parameters":[{"name":"policyId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LifecycleRequest"}}}},"responses":{"200":{"description":"OK"}}}},"/api/v1/policies/resolve":{"post":{"tags":["policy-controller"],"operationId":"resolvePolicy","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResolvePolicyRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string","format":"uuid"}}}}}}}},"/api/v1/identity/sync":{"post":{"tags":["identity-profile-controller"],"operationId":"syncProfile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdentitySyncRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/customers":{"get":{"tags":["customer-controller"],"operationId":"listCustomers","parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string"}},{"name":"customerType","in":"query","required":false,"schema":{"type":"string"}},{"name":"status","in":"query","required":false,"schema":{"type":"string"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageCustomerResponse"}}}}}},"post":{"tags":["customer-controller"],"operationId":"createCustomer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCustomerRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CustomerResponse"}}}}}}},"/api/v1/customers/{id}/lifecycle/deactivate":{"post":{"tags":["customer-controller"],"operationId":"deactivateCustomer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/customers/{id}/lifecycle/activate":{"post":{"tags":["customer-controller"],"operationId":"activateCustomer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/policies/{policyId}":{"get":{"tags":["policy-controller"],"operationId":"getPolicy","parameters":[{"name":"policyId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PolicyResponse"}}}}}},"patch":{"tags":["policy-controller"],"operationId":"updatePolicy","parameters":[{"name":"policyId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePolicyRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PolicyResponse"}}}}}}},"/api/v1/policies/{policyId}/premium":{"patch":{"tags":["policy-controller"],"operationId":"updatePremium","parameters":[{"name":"policyId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePremiumRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/reporting/policies/statistics":{"get":{"tags":["reporting-controller"],"operationId":"getPolicyStatistics","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PolicyStatisticsResponse"}}}}}}},"/api/v1/reporting/commissions/statistics":{"get":{"tags":["reporting-controller"],"operationId":"getCommissionStatistics","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommissionStatisticsResponse"}}}}}}},"/api/v1/hierarchy/dealers":{"get":{"tags":["organization-hierarchy-controller"],"operationId":"getDealers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HierarchyNodeResponse"}}}}}}}},"/api/v1/hierarchy/dealers/{dealerId}/branches":{"get":{"tags":["organization-hierarchy-controller"],"operationId":"getBranches","parameters":[{"name":"dealerId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HierarchyNodeResponse"}}}}}}}},"/api/v1/hierarchy/branches/{branchId}/agents":{"get":{"tags":["organization-hierarchy-controller"],"operationId":"getAgents","parameters":[{"name":"branchId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HierarchyNodeResponse"}}}}}}}}},"components":{"schemas":{"UpdateProductRequest":{"type":"object","properties":{"name":{"type":"string"},"category":{"type":"string"}}},"ProductResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"category":{"type":"string"},"status":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"version":{"type":"integer","format":"int64"}}},"ConfigureCommissionRequest":{"type":"object","properties":{"commissionType":{"type":"string"},"totalCommissionValue":{"type":"number"},"agentAShare":{"type":"number"},"agentBShare":{"type":"number"}}},"UpdateCustomerRequest":{"required":["addressInfo","contactInfo","name"],"type":"object","properties":{"name":{"type":"string"},"contactInfo":{"type":"string"},"addressInfo":{"type":"string"},"individualInfo":{"type":"string"},"businessInfo":{"type":"string"}}},"CustomerResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"customerType":{"type":"string"},"name":{"type":"string"},"contactInfo":{"type":"string"},"addressInfo":{"type":"string"},"status":{"type":"string"},"individualInfo":{"type":"string"},"businessInfo":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"CreateProductRequest":{"type":"object","properties":{"name":{"type":"string"},"category":{"type":"string"}}},"CreatePolicyRequest":{"type":"object","properties":{"policyNumber":{"type":"string"},"customerId":{"type":"string","format":"uuid"},"agentAId":{"type":"string","format":"uuid"},"agentBId":{"type":"string","format":"uuid"},"branchId":{"type":"string","format":"uuid"}}},"PolicyResponse":{"type":"object","properties":{"policyId":{"type":"string","format":"uuid"},"policyNumber":{"type":"string"},"customerId":{"type":"string","format":"uuid"},"agentAId":{"type":"string","format":"uuid"},"agentBId":{"type":"string","format":"uuid"},"branchId":{"type":"string","format":"uuid"},"premium":{"type":"number"},"status":{"type":"string"}}},"LifecycleRequest":{"type":"object","properties":{"isCommissionConfigured":{"type":"boolean"}}},"ResolvePolicyRequest":{"type":"object","properties":{"policyNumber":{"type":"string"}}},"IdentitySyncRequest":{"type":"object","properties":{"identityId":{"type":"string","format":"uuid"},"displayName":{"type":"string"},"email":{"type":"string"}}},"CreateCustomerRequest":{"required":["addressInfo","contactInfo","customerType","name"],"type":"object","properties":{"customerType":{"type":"string","enum":["INDIVIDUAL","ORGANIZATION"]},"name":{"type":"string"},"contactInfo":{"type":"string"},"addressInfo":{"type":"string"},"individualInfo":{"type":"string"},"businessInfo":{"type":"string"},"targetDealerId":{"type":"string","format":"uuid"},"targetBranchId":{"type":"string","format":"uuid"},"targetAgentId":{"type":"string","format":"uuid"}}},"UpdatePolicyRequest":{"type":"object","properties":{"customerId":{"type":"string","format":"uuid"},"agentAId":{"type":"string","format":"uuid"},"agentBId":{"type":"string","format":"uuid"},"branchId":{"type":"string","format":"uuid"}}},"UpdatePremiumRequest":{"type":"object","properties":{"premium":{"type":"number"}}},"PolicyStatisticsResponse":{"type":"object","properties":{"totalPolicies":{"type":"integer","format":"int64"},"draftCount":{"type":"integer","format":"int64"},"activeCount":{"type":"integer","format":"int64"},"inactiveCount":{"type":"integer","format":"int64"}}},"CommissionStatisticsResponse":{"type":"object","properties":{"totalCommissionAmount":{"type":"number"},"agentACommissionAmount":{"type":"number"},"agentBCommissionAmount":{"type":"number"},"configuredCommissionCount":{"type":"integer","format":"int64"}}},"Pageable":{"type":"object","properties":{"page":{"minimum":0,"type":"integer","format":"int32"},"size":{"minimum":1,"type":"integer","format":"int32"},"sort":{"type":"array","items":{"type":"string"}}}},"PageProductResponse":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/ProductResponse"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"paged":{"type":"boolean"},"unpaged":{"type":"boolean"},"pageSize":{"type":"integer","format":"int32"},"pageNumber":{"type":"integer","format":"int32"},"offset":{"type":"integer","format":"int64"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}}}},"SortObject":{"type":"object","properties":{"direction":{"type":"string"},"nullHandling":{"type":"string"},"ascending":{"type":"boolean"},"property":{"type":"string"},"ignoreCase":{"type":"boolean"}}},"PagePolicyResponse":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/PolicyResponse"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"empty":{"type":"boolean"}}},"HierarchyNodeResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"}}},"PageCustomerResponse":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/CustomerResponse"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"empty":{"type":"boolean"}}}}}}